Fix missed box edge click
Changes attempt to fix the missed box edge click when the box is
rotated. This fix will not work if there are large rotations in the ROI
box. We could error check this by finding if the box is where we expect
it to be. I will work on this.